Increase the initial stack from 16 to 24 bytes
This makes it possible to preserve the initial stack pointer without any
"dirty hacks", as explained in intro3_advanced.s. This is the most
convenient way to access the contents of the argument file (if provided).
Placing a label at the end of the source code is not reliable when
assembling a multi-file project.