Check for already defined labels
This would sometimes lead to infinite loops in the 'compose' method.
The new error is reported on the wrong line, and we should also have
unit tests for such error messages, but this will have to wait.

We also report an error if the label name is already an abbreviation.
However, an abbreviation may still hide an existing label or abbreviation.
Maybe this should also trigger an error?

This closes #5.