Document and improve initial code
The two usecases, either specifying the entry point or not, is now
properly documented in intro3_advanced.s. If the end stack has a single
element, then it is (truncated to 32 bits and) used as the exit status.
Otherwise, the status is 2. This does not apply to "ivm check", but the
exit status is always 1 if the machine crashes.

This closes #23.